+++ nodejs/fastWorker.js
@@ -28,8 +28,6 @@ const axios = require('axios');
 const Ajv = require('ajv');
 const merge = require('deepmerge');
 const Mustache = require('mustache');
-const JSONPath = require('jsonpath-plus').JSONPath;
-
 const semver = require('semver');
 
 const fast = require('@f5devcentral/f5-fast-core');
@@ -43,6 +41,7 @@ const DataStoreTemplateProvider = fast.DataStoreTemplateProvider;
 const StorageDataGroup = fast.dataStores.StorageDataGroup;
 const AS3Driver = drivers.AS3Driver;
 const TransactionLogger = fast.TransactionLogger;
+const IpamProviders = require('../lib/ipam');
 
 const pkg = require('../package.json');
 
@@ -61,17 +60,24 @@ if (typeof bigipStrictCert === 'string') {
     );
 }
 
-const ajv = new Ajv();
+const ajv = new Ajv({
+    useDefaults: true
+});
 ajv.addFormat('checkbox', /.*/);
 ajv.addFormat('table', /.*/);
 ajv.addFormat('password', /.*/);
 ajv.addFormat('text', /.*/);
 ajv.addFormat('grid-strict', /.*/);
 
+// Disable HTML escaping
+Mustache.escape = function escape(text) {
+    return text;
+};
+
 const configPath = process.AFL_TW_ROOT || `/var/config/rest/iapps/${projectName}`;
 const templatesPath = process.AFL_TW_TS || `${configPath}/templatesets`;
+const uploadPath = process.env.FAST_UPLOAD_DIR || '/var/config/rest/downloads';
 const scratchPath = `${configPath}/scratch`;
-const uploadPath = '/var/config/rest/downloads';
 const dataGroupPath = `/Common/${projectName}/dataStore`;
 
 const configDGPath = `/Common/${projectName}/config`;
@@ -79,6 +85,7 @@ const configKey = 'config';
 // Known good hashes for template sets
 const supportedHashes = {
     'bigip-fast-templates': [
+        '54ddc7113f372110fbe84e2730c2b60a7e641ba54e4201334929457540ebaeb7', // 1.10
         '89f6d8fb68435c93748de3f175f208714dcbd75de37d9286a923656971c939f0', // v1.9
         'fbaee3fd9ecce14a2d90df8c155998749b49126e0eb80267e9b426c58677a164', // v1.8.1
         '42650496f8e1b00a7e8e6a7c148a781bb4204e95f09f66d7d89af5793ae0b8b7', // v1.8
@@ -97,12 +104,15 @@ class FASTWorker {
         options = options || {};
         this.state = {};
 
+        this.baseUserAgent = `${pkg.name}/${pkg.version}`;
+        this.incomingUserAgent = '';
+
         this.isPublic = true;
         this.isPassThrough = true;
         this.WORKER_URI_PATH = `shared/${endpointName}`;
         this.driver = new AS3Driver({
             endPointUrl: `${bigipHost}/mgmt/shared/appsvcs`,
-            userAgent: `${pkg.name}/${pkg.version}`,
+            userAgent: this.baseUserAgent,
             bigipUser,
             bigipPassword,
             strictCerts: bigipStrictCert
@@ -127,6 +137,7 @@ class FASTWorker {
                 this.logger.fine(`FAST Worker [${id}]: ${text} took ${deltaTime}ms to complete`);
             }
         );
+        this.ipamProviders = options.ipamProviders;
 
         this.endpoint = axios.create({
             baseURL: bigipHost,
@@ -156,11 +167,27 @@ class FASTWorker {
         };
     }
 
+    validateConfig(config) {
+        return Promise.resolve()
+            .then(() => ajv.compile(this.getConfigSchema()))
+            .then((validate) => {
+                const valid = validate(config);
+                if (!valid) {
+                    return Promise.reject(new Error(
+                        `invalid config: ${validate.errors}`
+                    ));
+                }
+
+                return Promise.resolve(config);
+            });
+    }
+
     getConfig(reqid) {
         reqid = reqid || 0;
         const defaultConfig = {
             deletedTemplateSets: [],
-            ipamProviders: []
+            ipamProviders: [],
+            disableDeclarationCache: false
         };
         return Promise.resolve()
             .then(() => this.enterTransaction(reqid, 'gathering config data'))
@@ -195,93 +222,6 @@ class FASTWorker {
             });
     }
 
-    createIPAMProviderSchema(service, overrides) {
-        overrides = overrides || {};
-
-        return {
-            type: 'object',
-            title: service,
-            format: 'grid-strict',
-            properties: merge({
-                name: {
-                    title: 'Name',
-                    type: 'string',
-                    options: {
-                        grid_columns: 2,
-                        grid_break: true
-                    }
-                },
-                host: {
-                    title: 'Host',
-                    type: 'string',
-                    options: {
-                        grid_columns: 4
-                    }
-                },
-                username: {
-                    title: 'Username',
-                    type: 'string',
-                    options: {
-                        grid_columns: 4
-                    }
-                },
-                password: {
-                    title: 'Password',
-                    type: 'string',
-                    format: 'password',
-                    options: {
-                        grid_columns: 4,
-                        grid_break: true
-                    }
-                },
-                retrieveUrl: {
-                    title: 'Retrieve URL',
-                    type: 'string',
-                    format: 'text',
-                    options: {
-                        grid_columns: 4
-                    }
-                },
-                retrieveBody: {
-                    title: 'Retrieve Body',
-                    type: 'string',
-                    format: 'text',
-                    default: '{}',
-                    options: {
-                        grid_columns: 4
-                    }
-                },
-                retrievePathQuery: {
-                    title: 'Retrieve Path Query',
-                    type: 'string',
-                    format: 'text',
-                    default: '$',
-                    options: {
-                        grid_columns: 4,
-                        grid_break: true
-                    }
-                },
-                releaseUrl: {
-                    title: 'Release URL',
-                    type: 'string',
-                    format: 'text',
-                    options: {
-                        grid_columns: 4
-                    }
-                },
-                releaseBody: {
-                    title: 'Release Body',
-                    type: 'string',
-                    format: 'text',
-                    default: '{}',
-                    options: {
-                        grid_columns: 4
-                    }
-                }
-            }, overrides)
-        };
-    }
-
     getConfigSchema() {
         const baseSchema = {
             $schema: 'http://json-schema.org/schema#',
@@ -298,65 +238,24 @@ class FASTWorker {
                         hidden: true
                     }
                 },
+                disableDeclarationCache: {
+                    title: 'Disable AS3 Declaration Cache',
+                    type: 'boolean',
+                    description: [
+                        'Do not cache AS3 declarations.',
+                        'This ensures FAST is always using up-to-date declarations from AS3,',
+                        'which is only an issue if something other than FAST (e.g., config sync) is modifying AS3 config.',
+                        'Disabling declaration caching will negatively impact FAST performance.'
+                    ].join(' '),
+                    format: 'checkbox'
+                },
                 ipamProviders: {
                     title: 'IPAM Providers (Experimental/Beta)',
                     description: 'Configure IPAM providers that can be used in FAST templates to automatically manage IP addresses',
                     format: 'table',
                     type: 'array',
                     items: {
-                        anyOf: [
-                            // Infoblox support is untested
-                            // this.createIPAMProviderSchema('Infoblox', {
-                            //     apiVersion: {
-                            //         title: 'API Version',
-                            //         type: 'string',
-                            //         default: 'V2.11',
-                            //         options: {
-                            //             grid_columns: 2
-                            //         }
-                            //     },
-                            //     network: {
-                            //         title: 'Network Name',
-                            //         type: 'string',
-                            //         options: {
-                            //             grid_columns: 3,
-                            //             grid_break: true
-                            //         }
-                            //     },
-                            //     retrieveUrl: {
-                            // eslint-disable-next-line max-len
-                            //         const: '{{host}}/wapi/{{apiVersion}}/network/{{network}}?_function=next_available_ip&_return_as_object=1',
-                            //         options: {
-                            //             hidden: true
-                            //         }
-                            //     },
-                            //     retrieveBody: {
-                            //         const: '{ "num": 1 }',
-                            //         options: {
-                            //             hidden: true
-                            //         }
-                            //     },
-                            //     retrievePathQuery: {
-                            //         const: '$.ipv4addrs[0].ipv4addr',
-                            //         options: {
-                            //             hidden: true
-                            //         }
-                            //     },
-                            //     releaseUrl: {
-                            //         const: '{{host}}/wapi/{{apiVersion}}/ipv4address/{{network}}:{{addr}}',
-                            //         options: {
-                            //             hidden: true
-                            //         }
-                            //     },
-                            //     releaseBody: {
-                            //         const: '{}',
-                            //         options: {
-                            //             hidden: true
-                            //         }
-                            //     }
-                            // }),
-                            this.createIPAMProviderSchema('Generic')
-                        ]
+                        anyOf: this.ipamProviders.getSchemas()
                     }
                 }
             },
@@ -430,6 +329,12 @@ class FASTWorker {
      */
     onStart(success, error) {
         this.hookCompleteRestOp();
+        // instantiate here to ensure logger instance is ready
+        this.ipamProviders = new IpamProviders({
+            secretsManager: this.secretsManager,
+            transactionLogger: this.transactionLogger,
+            logger: this.logger
+        });
         this.logger.fine(`FAST Worker: Starting ${pkg.name} v${pkg.version}`);
         this.logger.fine(`FAST Worker: Targetting ${bigipHost}`);
         const startTime = Date.now();
@@ -548,7 +453,10 @@ class FASTWorker {
      */
     sendTeemReport(reportName, reportVersion, data) {
         const documentName = `${projectName}: ${reportName}`;
-        return this.teemDevice.report(documentName, `${reportVersion}`, {}, data)
+        const baseData = {
+            userAgent: this.incomingUserAgent
+        };
+        return this.teemDevice.report(documentName, `${reportVersion}`, baseData, data)
             .catch(e => this.logger.error(`FAST Worker failed to send telemetry data: ${e.stack}`));
     }
 
@@ -867,8 +775,15 @@ class FASTWorker {
                 if (value[childName]) {
                     acc[key] = value;
                 }
-                if (value.items && value.items[childName]) {
-                    acc[`${key}.items`] = value.items;
+                if (value.items) {
+                    if (value.items[childName]) {
+                        acc[`${key}.items`] = value.items;
+                    } else if (value.items.oneOf) {
+                        const prop = value.items.oneOf.find(i => i[childName]);
+                        if (typeof prop !== 'undefined') {
+                            acc[key] = prop;
+                        }
+                    }
                 }
                 return acc;
             }, {});
@@ -928,38 +843,41 @@ class FASTWorker {
                 });
             })
             .then(() => Promise.all(Object.values(enumFromBigipProps).map((prop) => {
-                const endPoint = `/mgmt/tm/${prop.enumFromBigip}?$select=fullPath`;
+                const epStubs = Array.isArray(prop.enumFromBigip) ? prop.enumFromBigip : [prop.enumFromBigip];
+                const endPoints = epStubs.map(x => `/mgmt/tm/${x}?$select=fullPath`);
                 return Promise.resolve()
-                    .then(() => {
-                        if (this._hydrateCache[endPoint]) {
-                            return this._hydrateCache[endPoint];
-                        }
+                    .then(() => Promise.all(endPoints.map(endPoint => Promise.resolve()
+                        .then(() => {
+                            if (this._hydrateCache[endPoint]) {
+                                return this._hydrateCache[endPoint];
+                            }
 
-                        return this.recordTransaction(
-                            requestId, `fetching data from ${endPoint}`,
-                            this.endpoint.get(endPoint)
-                        )
-                            .then((response) => {
-                                const items = response.data.items;
-                                this._hydrateCache[endPoint] = items;
-                                return items;
-                            });
-                    })
-                    .then((items) => {
-                        if (items) {
-                            return Promise.resolve(items.map(x => x.fullPath));
-                        }
-                        return Promise.resolve([]);
-                    })
-                    .catch(e => this.handleResponseError(e, `GET to ${endPoint}`))
+                            return this.recordTransaction(
+                                requestId, `fetching data from ${endPoint}`,
+                                this.endpoint.get(endPoint)
+                            )
+                                .then((response) => {
+                                    const items = response.data.items;
+                                    this._hydrateCache[endPoint] = items;
+                                    return items;
+                                });
+                        })
+                        .then((items) => {
+                            if (items) {
+                                return Promise.resolve(items.map(x => x.fullPath));
+                            }
+                            return Promise.resolve([]);
+                        })
+                        .catch(e => this.handleResponseError(e, `GET to ${endPoint}`))
+                        .catch(e => Promise.reject(new Error(`Failed to hydrate ${endPoint}\n${e.stack}`))))))
+                    .then(itemsArrays => itemsArrays.flat())
                     .then((items) => {
                         if (items.length !== 0) {
                             prop.enum = items;
                         } else {
                             prop.enum = [null];
                         }
-                    })
-                    .catch(e => Promise.reject(new Error(`Failed to hydrate ${endPoint}\n${e.stack}`)));
+                    });
             })))
             .then(() => schema);
     }
@@ -1016,89 +934,6 @@ class FASTWorker {
             .then(() => apps);
     }
 
-    populateIPAMAddress(template, templateData, config, reqid, ipamAddrs) {
-        let ipamChain = Promise.resolve();
-        const schema = template.getParametersSchema();
-        const ipFromIpamProps = this.getPropsWithChild(schema, 'ipFromIpam');
-        Object.entries(ipFromIpamProps).forEach(([name, prop]) => {
-            const providerName = templateData.parameters[name];
-            const provider = config.ipamProviders
-                .find(p => p.name === providerName);
-            if (provider) {
-                delete prop.enum;
-                if (!ipamAddrs[providerName]) {
-                    ipamAddrs[providerName] = [];
-                }
-                ipamChain = ipamChain
-                    .then(() => this.secretsManager.decrypt(provider.password))
-                    .then(providerPassword => this.recordTransaction(
-                        reqid, `fetching address from IPAM provider: ${providerName}`,
-                        axios.post(
-                            Mustache.render(provider.retrieveUrl, provider),
-                            JSON.parse(Mustache.render(provider.retrieveBody, provider)),
-                            {
-                                auth: {
-                                    username: provider.username,
-                                    password: providerPassword
-                                }
-                            }
-                        )
-                    ))
-                    .catch(e => Promise.reject(new Error(
-                        `failed to get IP address from IPAM provider (${providerName}): ${e.stack}`
-                    )))
-                    .then((res) => {
-                        let value = '';
-                        try {
-                            value = JSON.parse(res.data);
-                        } catch (e) {
-                            value = res.data;
-                        }
-                        value = JSONPath(provider.retrievePathQuery, value)[0];
-                        templateData.parameters[name] = value;
-                        ipamAddrs[providerName].push(value);
-                    });
-            }
-        });
-        return ipamChain;
-    }
-
-    releaseIPAMAddress(reqid, config, appData, excludeAddrs) {
-        if (!appData.ipamAddrs) {
-            return Promise.resolve();
-        }
-        const promises = [];
-        Object.entries(appData.ipamAddrs).forEach(([providerName, addrs]) => {
-            const provider = config.ipamProviders
-                .filter(p => p.name === providerName)[0];
-            addrs.forEach((address) => {
-                if (!excludeAddrs || !excludeAddrs[provider] || !excludeAddrs[provider].find(a => a === address)) {
-                    const view = Object.assign({}, provider, { address });
-                    promises.push(Promise.resolve()
-                        .then(() => this.secretsManager.decrypt(provider.password))
-                        .then(providerPassword => this.recordTransaction(
-                            reqid, `releasing ${address} from IPAM provider: ${providerName}`,
-                            axios.post(
-                                Mustache.render(provider.releaseUrl, view),
-                                JSON.parse(Mustache.render(provider.releaseBody, view)),
-                                {
-                                    auth: {
-                                        username: provider.username,
-                                        password: providerPassword
-                                    }
-                                }
-                            )
-                        ))
-                        .catch(e => this.logger.error(
-                            `failed to release IP address from IPAM provider (${providerName}): ${e.stack}`
-                        )));
-                }
-            });
-        });
-
-        return Promise.all(promises);
-    }
-
     releaseIPAMAddressesFromApps(reqid, appsData) {
         let config;
         const promises = [];
@@ -1123,7 +958,7 @@ class FASTWorker {
                     return this.getConfig(reqid)
                         .then((c) => { config = c; });
                 })
-                .then(() => this.releaseIPAMAddress(reqid, config, view)));
+                .then(() => this.ipamProviders.releaseIPAMAddress(reqid, config, view)));
         });
 
         return Promise.all(promises);
@@ -1171,8 +1006,12 @@ class FASTWorker {
                     this.templateProvider.fetch(tmplData.name)
                 ))
                 .catch(e => Promise.reject(new Error(`unable to load template: ${tmplData.name}\n${e.stack}`)))
-                .then(tmpl => this.populateIPAMAddress(tmpl, tmplData, config, reqid, ipamAddrs)
-                    .then(() => tmpl))
+                .then((tmpl) => {
+                    const schema = tmpl.getParametersSchema();
+                    const ipFromIpamProps = this.getPropsWithChild(schema, 'ipFromIpam');
+                    return this.ipamProviders.populateIPAMAddress(ipFromIpamProps, tmplData, config, reqid, ipamAddrs)
+                        .then(() => tmpl);
+                })
                 .then(tmpl => this.recordTransaction(
                     reqid, `rendering template (${tmplData.name})`,
                     tmpl.fetchAndRender(tmplData.parameters)
@@ -1180,7 +1019,7 @@ class FASTWorker {
                 .then(rendered => JSON.parse(rendered))
                 .catch(e => Promise.resolve()
                     // Release any IPAM IP addrs
-                    .then(() => this.releaseIPAMAddress(reqid, config, { ipamAddrs }))
+                    .then(() => this.ipamProviders.releaseIPAMAddress(reqid, config, { ipamAddrs }))
                     // Now re-reject
                     .then(() => Promise.reject(new Error(`failed to render template: ${tmplData.name}\n${e.stack}`))))
                 .then((decl) => {
@@ -1198,7 +1037,7 @@ class FASTWorker {
 
                     const oldAppData = tmplData.previousDef || {};
                     if (oldAppData.ipamAddrs) {
-                        this.releaseIPAMAddress(reqid, config, oldAppData, ipamAddrs);
+                        this.ipamProviders.releaseIPAMAddress(reqid, config, oldAppData, ipamAddrs);
                     }
                 });
         });
@@ -1213,9 +1052,17 @@ class FASTWorker {
      * HTTP/REST handlers
      */
     recordRestRequest(restOp) {
+        // Update driver's user agent if one was provided with the request
+        const userAgent = restOp.getUri().query.userAgent;
+        this.incomingUserAgent = userAgent || '';
+        this.driver.userAgent = userAgent ? `${userAgent};${this.baseUserAgent}` : this.baseUserAgent;
+
+        // Record the time we received the request
         this.requestTimes[restOp.requestId] = Date.now();
+
+        // Dump information to the log
         this.logger.fine(
-            `FAST Worker [${restOp.requestId}]: received request method=${restOp.getMethod()}; path=${restOp.getUri().pathname}`
+            `FAST Worker [${restOp.requestId}]: received request method=${restOp.getMethod()}; path=${restOp.getUri().pathname}; userAgent=${this.incomingUserAgent}`
         );
     }
 
@@ -1590,7 +1437,7 @@ class FASTWorker {
                 reqid, 'validate template set',
                 this._validateTemplateSet(scratchPath)
             ))
-            .catch(e => Promise.reject(new Error(`Template set (${tsid}) failed validation: ${e.message}`)))
+            .catch(e => Promise.reject(new Error(`Template set (${tsid}) failed validation: ${e.message}. ${e.stack}`)))
             .then(() => this.enterTransaction(reqid, 'write new template set to data store'))
             .then(() => this.templateProvider.invalidateCache())
             .then(() => DataStoreTemplateProvider.fromFs(this.storage, scratchPath, [tsid]))
@@ -1633,18 +1480,11 @@ class FASTWorker {
         const reqid = restOperation.requestId;
 
         return Promise.resolve()
-            .then(() => ajv.compile(this.getConfigSchema()))
-            .then((validate) => {
-                const valid = validate(config);
-                if (!valid) {
-                    return Promise.reject(this.genRestResponse(
-                        restOperation, 422,
-                        `supplied settings were not valid:\n${validate.errors}`
-                    ));
-                }
-
-                return Promise.resolve();
-            })
+            .then(() => this.validateConfig(config))
+            .catch(e => Promise.reject(this.genRestResponse(
+                restOperation, 422,
+                `supplied settings were not valid:\n${e.message}`
+            )))
             .then(() => this.getConfig(reqid))
             .then(prevConfig => this.encryptConfigSecrets(config, prevConfig))
             .then(() => this.gatherProvisionData(reqid, true))
@@ -1915,18 +1755,11 @@ class FASTWorker {
             .then((prevConfig) => {
                 combinedConfig = Object.assign({}, prevConfig, config);
             })
-            .then(() => ajv.compile(this.getConfigSchema()))
-            .then((validate) => {
-                const valid = validate(combinedConfig);
-                if (!valid) {
-                    return Promise.reject(this.genRestResponse(
-                        restOperation, 422,
-                        `supplied settings were not valid:\n${validate.errors}`
-                    ));
-                }
-
-                return Promise.resolve();
-            })
+            .then(() => this.validateConfig(combinedConfig))
+            .catch(e => Promise.reject(this.genRestResponse(
+                restOperation, 422,
+                `supplied settings were not valid:\n${e.message}`
+            )))
             .then(() => this.gatherProvisionData(reqid, true))
             .then(provisionData => this.driver.setSettings(combinedConfig, provisionData))
             .then(() => this.saveConfig(combinedConfig, reqid))