+++ test/integ/integ.js
@@ -73,9 +73,92 @@ function getAuthToken() {
         .catch(err => Promise.reject(new Error(`Unable to generate auth token: ${err.message}`)));
 }
 
+function deleteAllApplications() {
+    return Promise.resolve()
+        .then(() => endpoint.delete('/mgmt/shared/fast/applications'))
+        .then((response) => {
+            const taskid = response.data.id;
+            if (!taskid) {
+                console.log(response.data);
+                assert(false, 'failed to get a taskid');
+            }
+            return waitForCompletedTask(taskid);
+        })
+        .then((task) => {
+            if (task.code !== 200) {
+                console.log(task);
+            }
+            assert.strictEqual(task.code, 200);
+        })
+        .catch(err => Promise.reject(new Error(`Failed to delete applications: ${err.message}`)));
+}
+
+
+describe('Template Sets', function () {
+    this.timeout(120000);
+    const url = '/mgmt/shared/fast/templatesets';
+
+    function assertGet(expected, templateSetId) {
+        const fullUrl = templateSetId ? `${url}/${templateSetId}` : url;
+        return Promise.resolve()
+            .then(() => endpoint.get(fullUrl))
+            .then((actual) => {
+                assert.strictEqual(actual.status, expected.status);
+                if (Array.isArray(expected.data)) {
+                    expected.data.forEach((expTemplateSet) => {
+                        const templateSet = templateSetId ? actual.data
+                            : actual.data.find(ts => ts.name === expTemplateSet.name);
+                        assert.ok(templateSet, `Template with name ${expTemplateSet.name} must exist`);
+                        assert.ok(templateSet.templates.length > 0, 'Template set must contain templates');
+                        assert.ok(templateSet.hash, 'Template set must contain a hash');
+                        assert.strictEqual(templateSet.supported, expTemplateSet.supported, 'Template set must show correct supported flag');
+                    });
+                } else {
+                    assert.deepStrictEqual(actual.data, expected.data);
+                }
+            })
+            .catch((e) => {
+                if (e.response && !expected.error) {
+                    console.error(e.response.data);
+                    return Promise.reject(e);
+                }
+                assert.deepStrictEqual(e.response.data, expected.error);
+                assert.strictEqual(e.response.status, expected.status);
+                return Promise.resolve();
+            });
+    }
+
+    before(() => getAuthToken());
+    before('Delete all applications', deleteAllApplications);
+
+    it('GET built-in template sets', () => Promise.resolve()
+        .then(() => assertGet({
+            data: [
+                { name: 'bigip-fast-templates', supported: true },
+                { name: 'examples', supported: false }
+            ],
+            status: 200
+        })));
+    it('DELETE template set by ID', () => Promise.resolve()
+        .then(() => endpoint.delete(`${url}/examples`))
+        .then((actual) => {
+            assert.strictEqual(actual.status, 200);
+            assert.deepStrictEqual(actual.data, { code: 200, message: 'success' });
+            return assertGet({ error: { code: 404, message: 'Template set examples does not exist' }, status: 404 }, 'examples');
+        }));
+    it('POST re-install template set and GET by ID', () => Promise.resolve()
+        .then(() => endpoint.post(url, { name: 'examples' }))
+        .then((actual) => {
+            assert.strictEqual(actual.status, 200);
+            assert.deepStrictEqual(actual.data, { code: 200, message: '' });
+            return assertGet({ data: [{ name: 'examples', supported: false }], status: 200 }, 'examples');
+        }));
+});
+
 describe('Applications', function () {
     this.timeout(120000);
     before(() => getAuthToken());
+    before('Delete all applications', deleteAllApplications);
 
     function deployApplication(templateName, parameters) {
         parameters = parameters || {};
@@ -107,24 +190,6 @@ describe('Applications', function () {
             });
     }
 
-    before('Delete all applications', () => Promise.resolve()
-        .then(() => endpoint.delete('/mgmt/shared/fast/applications'))
-        .then((response) => {
-            const taskid = response.data.id;
-            if (!taskid) {
-                console.log(response.data);
-                assert(false, 'failed to get a taskid');
-            }
-            return waitForCompletedTask(taskid);
-        })
-        .then((task) => {
-            if (task.code !== 200) {
-                console.log(task);
-            }
-            assert.strictEqual(task.code, 200);
-        })
-        .catch(err => Promise.reject(new Error(`Failed to delete applications: ${err.message}`))));
-
     it('Deploy examples/simple_udp_defaults', () => deployApplication('examples/simple_udp_defaults'));
 
     it('Deploy bigip-fast-templates/http', () => deployApplication('bigip-fast-templates/http', {
@@ -132,7 +197,7 @@ describe('Applications', function () {
         app_name: 'HTTP_App',
         virtual_address: '10.0.0.1',
         pool_members: [
-            { serverAddresses: ['10.0.0.1'], servicePort: 80 }
+            { serverAddresses: ['10.0.0.1'] }
         ]
     }));
     it('Deploy bigip-fast-templates/tcp', () => deployApplication('bigip-fast-templates/tcp', {
@@ -140,7 +205,7 @@ describe('Applications', function () {
         app_name: 'TCP-App',
         virtual_address: '10.0.0.2',
         pool_members: [
-            { serverAddresses: ['10.0.0.2'], servicePort: 80 }
+            { serverAddresses: ['10.0.0.2'] }
         ]
     }));
     it('Deploy bigip-fast-templates/dns', () => deployApplication('bigip-fast-templates/dns', {
@@ -148,7 +213,7 @@ describe('Applications', function () {
         app_name: 'DNS-App',
         virtual_address: '10.0.0.3',
         pool_members: [
-            { serverAddresses: ['10.0.0.3'], servicePort: 80 }
+            { serverAddresses: ['10.0.0.3'] }
         ],
         monitor_queryName: 'example.com'
     }));
@@ -164,7 +229,7 @@ describe('Applications', function () {
         tenant_name: 't-5',
         virtual_address: '10.0.0.5',
         pool_members: [
-            { serverAddresses: ['10.0.0.5'], servicePort: 80 }
+            { serverAddresses: ['10.0.0.5'] }
         ],
         app_fqdn: 'example.com',
         monitor_fqdn: 'example.com'
@@ -174,7 +239,7 @@ describe('Applications', function () {
         app_name: 'LDAP-App',
         virtual_address: '10.0.0.6',
         pool_members: [
-            { serverAddresses: ['10.0.0.6'], servicePort: 80 }
+            { serverAddresses: ['10.0.0.6'] }
         ],
         monitor_passphrase: 'aa'
     }));
@@ -183,7 +248,7 @@ describe('Applications', function () {
         app_name: 'IIS_App',
         virtual_address: '10.0.0.7',
         pool_members: [
-            { serverAddresses: ['10.0.0.7'], servicePort: 80 }
+            { serverAddresses: ['10.0.0.7'] }
         ]
     }));
     it('Deploy bigip-fast-templates/smtp', () => deployApplication('bigip-fast-templates/smtp', {
@@ -191,7 +256,7 @@ describe('Applications', function () {
         app_name: 'SMTP-App',
         virtual_address: '10.0.0.8',
         pool_members: [
-            { serverAddresses: ['10.0.0.8'], servicePort: 80 }
+            { serverAddresses: ['10.0.0.8'] }
         ],
         monitor_domain: 'example.com'
     }));
@@ -234,7 +299,8 @@ describe('Settings', function () {
         .then(actual => assertResponse(actual, {
             data: {
                 deletedTemplateSets: [],
-                ipamProviders: []
+                ipamProviders: [],
+                disableDeclarationCache: false
             },
             status: 200
         })));
@@ -256,7 +322,8 @@ describe('Settings', function () {
                 network: 'testnetwork'
             }],
             log_afm: false,
-            log_asm: false
+            log_asm: false,
+            disableDeclarationCache: false
         };
         const expected = {
             data: { code: 200, message: '' },
@@ -291,7 +358,8 @@ describe('Settings', function () {
                     deletedTemplateSets: ['examples'],
                     ipamProviders: [],
                     log_afm: false,
-                    log_asm: false
+                    log_asm: false,
+                    disableDeclarationCache: false
                 };
                 return endpoint.get(url)
                     .then(res => assertResponse(res, expected));